// src/codegen/types.ts
export type Duration = {
  height: number;
} | {
  time: number;
};